+++ docs/src/changelog.md
@@ -12,6 +12,12 @@ If your currently installed *Quality-time* version is not the latest version, pl
 
 <!-- The line "## <square-bracket>Unreleased</square-bracket>" is replaced by the release/release.py script with the new release version and release date. -->
 
+## [Unreleased]
+
+### Fixed
+
+- Measuring source up-to-dateness of files stored in GitLab >= v17.7 would fail with Quality-time reporting a connection error (status code 404 and message "Not found"). Fixes [#11209](https://github.com/ICTU/quality-time/issues/11209).
+
 ## v5.29.0 - 2025-05-08
 
 ### Added